[main][execve] Run /sbin/init after init drivers
This relatively complex change runs /bin/init (or whatever program passed
in the cmdline) from the init drivers kthread, instead of from kmain().
The idea of running as `init` directly in kmain() turned out to be bad on
multiple levels. The hack of running it in SLEEP mode did not really solve
the problem because it still forced the loading init's ELF binary in kmain
while preemption was still disabled. And that is a problem because we want
to enforce all VFS calls to run with preemption enabled: think about the
case when the VFS wants to put the current task to sleep in order to wait
for I/O to complete. Therefore, the most flexible solution was just to run
`init` only when the kernel is fully initialized.

BONUS: execve() has been refactored again.