[common] Make ASSERT-like macros to turn into no-op blocks when disabled
Having a macro like ASSERT(x) that when disabled (non-DEBUG builds) is
defined just as:

	#define ASSERT(x)

Is error-prone. Consider the following code:

	if (something())
		ASSERT(x > 3);

	do_action();

In debug builds, everything will work as expected, but in release builds
the ASSERT(x > 3) will completely disappear, making the compiler to see
the following C code:

	if (something())
		do_action();

Which is *not* what we intended. That's why when disabled, macros like
ASSERT are now defined as:

	#define ASSERT(x) do { /* nothing */ } while (0)

This completely solves the problem.