[common][basic_defs] Drop the ugly `rshift()` inline function
That function was used during the attempt to make Tilck build with
-Wconversion even with GCC. Because of a compiler's bug, it was necessary
to wrap the right-shift operation in a always-inline function.
See https://stackoverflow.com/questions/54421737 for more details.

But after it turned out that GCC's -Wconversion flag is unusable in
practice because it requires casts even when implicit conversions occur
like this:

	u8 a=1, b=2, c;
	c = a + b;        /* GCC won't accept this with -Wconversion */
	c = (u8) (a + b); /* That's what GCC would accept */

I decided that it makes sense to compile only with clang's -Wconversion.
Setting CMake's WCONV=1 currently produces a warning in case the compiler
used for the kernel is not clang.

Anyway, it's worth dropping this ugly rshift() hack, now that's clear it
won't be used.