[vfs] Add the `res_last_sl` param to vfs_stat64() and vfs_resolve()
The `res_last_sl` tells if the caller wants the last component to be
resolved when it's a symlink or not. The parameter is currently ignored
by the only function that should understand it (vfs_resolve), but it's
propagated correctly to it (e.g. starting from stat and lstat).

In following changes, after symlinks have been introduced to ramfs, it will
be actually used in vfs_resolve().