[execve] Rename abs_path with just `path` everywhere
As part of the plan for supporting symlinks, the compute_abs_path() func
will have to die and the vfs_resolve() function will be the only one able
to both resolve symlinks and handle '.', '..', '//', '///' etc. The nice
hack of fixing the path upfront cannot solve the whole problem if symlinks
exist.

Therefore, the term 'abs_path' won't be correct anymore after the removal
of compute_abs_path(), because the path will be (potentially) relative and
contain symlinks as well.