[cmdline] Introduce the -sat option [sched alive thread]
This is a nice debug feature making the kernel to run a debug kthread
printing every second a message and the value of the counter. This is
useful in CI testing to check that, at least the scheduler and the timer
continued in work (or not) in case of a stall.