[debug][cov] Introduce set_sched_alive_thread_enabled()
When the sched alive thread is active, a message is printed each second
on the console by the kernel. That's fine while tests are running, but it's
painful when the devshell has to dump to the console the GCOV gcda files
in base64. Therefore, a function and a relative Tilck cmd has been created
in order to allow user apps to disable the printing of that message.