[system_tests] Fix cmd_fs3()
The test was trying to remove a file never created, expecting the operation
to fail. It did fail every time as expected, but that's was not the point
of the test: it had to try removing a file that has been recently created
and then removed while iterating over its parent directory.