[vfs] Make vfs_resolve() accept a char *last_comp parameter
This is a necessary step in order to allow vfs_resolve() to accept also
relative paths. In that case, at least for now, a normalized path will be
passed to __vfs_resolve() which sets rp->last_comp using the newer path
as reference. Therefore, rp->last_comp needs to point to buffer living
outside of vfs_resolve() in order to contain the last component of the
normalized path. NOTE: this is, very likely, *not* the definitive solution
for `last_comp`, but just a necessary incremental step.