[vfs] Make vfs_resolve() call compute_abs_path() internally
This is a first major step allowing the VFS layer to accept relative paths.
Clearly, the whole computation-of-the-absolute-path thing is still a
temporary solution good only when there are no symlinks. Still, it is good
to move that call down to vfs_resolve() in order to remove it from all the
fs-syscalls calling it before using any VFS function.