[vfs][resolve] Almost re-write the new resolve logic to stack of vfs_path
Until now, the new resolve function used separate stacks like `idir`, `fss`
and `pc_offs` and kept for the whole time a pointer to the resolved vfs_path
as ctx->rp. With this approach, even for a single-fs path, handling '..' was
a nightmare: it required going to the element before the previous and then
getting back the values for *rp of the previous element using get_entry().
It was already dirty and barely acceptable, but when it came to support '..'
crossing a mount-point boundary, it got way too much.

This change, greatly simplifies the whole implementation, by removing
pc_offs and the other separate stacks, along with the explicit `rp`. Now,
the caller of __vfs_resolve() just has to push on its stack what previously
was `rp`. Since now the stack is made by vfs_path elements, it's super-simple.
The final `rp` that we can think as "returned" by __vfs_resolve() is just
the vfs_path at position ctx->paths[ctx->ss - 1] (top of the stack).