[vfs] Start using pi->cwd2 in an unsafe way in vfs_resolve()
NOTE: unsafe means that all the necessary retain/release on the `fs` and
on the current working directory's inode is MISSING. That's OK as a first
step, but definitively it's not acceptable.

** BUGS **
With the current commit, a major bug gets introduced in vfs_resolve():
in case of relative paths, `..` does not work as it should. For example:

/# cd dev
/dev# ls
/dev# cd ..
/# ls

the last `ls` shows the same content as the one in `/dev`.