[vfs][resolve] Stop pushing on the stack when it's not necessary
With the new implementation of __vfs_res_handle_dot_dot(), the only
function that really needed the stack until now, being made to use just
get_entry(".."), it's possible now to just replace the last element of the
stack, instead of pushing on it. This means the whole stack can just be
discard again, BUT, let's keep it for the moment, given that it might be
strictly necessary for resolving symlinks.