[build_toolchain] Make `tcc` to be compiled with gcc+musl on i386
This solves the problem with the `tcc` unable to run on Tilck, when
everything is compiled under a i386 host machine (considered exotic
scenario). The problem was just that `tcc` ended up being linked
with glibc, which is not supported by Tilck. Now `tcc` gets linked
with libmusl even in this case and works on Tilck.