[init] Push back again init in /bin, dropping /sbin
After some evalutation, it turns out that it's just a waste to keep a root
level directory (/sbin) just for one application. Anyway, Tilck does not
support other users than root. Therefore, it's more convenient to just
print an error in case `init` is run not as a first process and put it back
in /init, making overall Tilck base filesystem simpler.