[fat32] Handle '.' and '..' in /
Unfortunately, fat32 is the only FS not having '.' and '..' in the root
directory, while assuming that each FS always has '.' and '..' in each
directory, makes the resolve code simpler and more elegant. Therefore, it's
worth adding a special mini-hack in fat_get_entry() to support '.' and '..'
even in the root directory.