[init] Add the -ns (no start) option
This option will allow the system tests to save some CPU time by just
passing this option to init. After all, the start script does nothing more
than setting up all the symlinks from / to /initrd in order to the system
be useable by an user and regular programs. Our devshell instead, does not
need that environment in order to run it's tests. The only thing it needs
is the presence of the directory /tmp, at the moment created by the kernel
itself.