[vfs][resolve] Always replace the stack top before return
Even in case rc != 0, __vfs_resolve() must replace the stack top with the
value of `np` because of the locking invariants of the function. Otherwise,
in case of some failures, the main vfs_resolve() function won't get the
right `vfs_path` and the right `fs` to unlock.

Example:
	/usr/bin -> /initrd/usr/bin/fb_drawing
	ln -s /initrd/usr/bin/fb_hello_world /usr/bin/

Now, the target path is a symlink that gets actually resolved, but it's not
a directory. Therefore vfs_resolve_have_to_stop() breaks the loop in
__vfs_resolve() and we check if rc == 0 before replacing the stack's top
with `np`. But the problem with that is that stack's top points still at
/usr/bin owned by ramfs, which has already been unlocked while resolving
the symlink: there's no turning back. Now we have to unlock whatever was
been locked last, which is the FS of /initrd/usr/bin/fb_drawing, fat.