[term] Allow enabling/disabling of the cursor
Until now, only term itself during scroll could disable the
cursor, temporarily. There was no such thing as "the cursor is
permanently disabled on this TTY". Now, the `term` layer supports
that, but the TTY still doesn't. The next step will be to add
support for that in TTY as well (using ^[[?25h and ^[[?25l).