[fb_console] Not using fast funcs on low-mem systems
The fast char-rendering functions require the pre-rendering of all the possible
scanlines for all the possible characters, for all the possible fg/bg color
combinations. That takes about 2 MB of memory and it's not worth loosing that
amount of memory if the system has less than 16 MB of free heap on boot (that
means having 16 MB + 1 MB (kernel) + 2-3 MB (fatpart) of physical memory).