[kmalloc] Remove metadata from chunks in small-heaps
Reason: there are in practice too few small heaps to justify the need for
a small heap pointer (+size, +align) in each chunk as metadata. The existence
of that metadata actually causes a fair amount of waste and makes us loose the
automatic alignment. That's why the align field has been added in small heap's
metadata and the 4 supported kmalloc align flags has been introduced, but it's
a lot of trouble, caused as a ripple-effect by the single decision to avoid
iterating over the list of small heaps on kfree2(). With this patch, the memory
is used in a more efficient way and the peak of small-heaps decreased from 27
to 17, after `runall`. Also the heap memory used on a `contrib` build, running
in text mode went down from 1044 KB to 976 KB. It's seems a remarkable
improvement.