[kmalloc] Make small heap's min block size = 32
After looking at the data thanks to kmalloc's "heavy stats" and the debug
panel, it was absolutely clear that there are almost no allocations with
size <= 16 byte. Therefore, reducing the size of small heaps' metadata by a
half and reducing by a half the depth of small heap's BST is totally worth.
The performance gain for that is also remarkable: from 350+ cycles for allocs
in the range 128-1024 to barely 304 at most with a minimum at 280 cycles.