[basic_defs] Intr. `offt` as a replacement of `off_t`
Currently, off_t is used internally in the kernel, even if it's size is not
"fixed" (pointer-size) for a given arch, but it's entirely dependent on
the headers exporting it. A long as it isn't used for syscalls, that's not
a problem, but it's an unsafe practice to keep a type like that used in the
kernel. Also, it makes sense to be sure about the size of each type inside
the kernel and not depend on anything else. On the other side, copy-pasting
all the defines and structs of types used as part of the Linux user/kernel
interface is annoying. Tilck cannot change those structs is any way and no
libc is supposed to support explicitly Tilck as kernel. Therefore, for structs
which layout cannot change there are still enough valid reasons to use "extern"
definitions from "system" includes, instead of having Tilck's own copy for each
one of them. But for off_t, that's wrong because size it's not fixed: it can
only be used as interface between user code and libc, not between libc and the
kernel. Therefore it must be banned from Tilck.