[config] Make the default fg color to be `white` (light gray)
This is kind-of useful just in order to allow some text to be "bold" which is
displayed by using the bright white color. Otherwise, the default color is
the same as the bold color and the visual effect is lost.

In addition, this change adds an option in the main CMake file in order to
allow users to choose at build time whether the default text color should be
bright white or not.