[build_toolchain] Make the busybox.config simpler
This is a necessary fix in order to support systems where, for some reason,
busybox's .config file gets re-ordered. Having the saved .config file to be
equal at binary level with the busybox .config file is necessary in order to
allow the check for that in the CMake files to pass. That check is useful to
force users to rebuild busybox in case its configuration is different from the
one expected by Tilck.