[paging] Introduce paging_hw.h
The purpose of this change is to reduce the number of files rebuilt when
some of the arch-specific (hal) headers change, by removing the include of
the master hal.h from paging.h (widely used header).

Now paging.h includes just hal_types.h which is nothing more than just a
bunch of forward-declarations, while the inline functions in paging which
required the inclusion of <hal.h> have been moved to paging_hw.h. The new
header is, of course, needed in much less places than paging.h.