[vfs] Fix broken gtests build on newer gcc compilers
In the vfs_path_funcs_wrapper() macro, the param `func` is brutally casted to
`vfs_func_impl`, which has a generic type (fs, vfs_path, uptr, uptr, uptr) and
it is _expected_ to be different than the type of `func`. But, the newest
GCC compilers, produce a -Werror because of that. Solution: first cast the
function pointer `func` to `void *` and then to `vfs_func_impl`.