[signal] Treat all custom signal handlers as SIG_IGN
Note: previously SIGINT was treated as SIG_DFL (default action) when it had
a custom handler. Now, because of ASH's implementation, in order to Ctrl+C
to behave as expected in most of the cases, like killing `cat` with Ctrl+C,
without making the whole ASH shell to die, it's necessary ignore SIGINT when
handled by the user code, instead of killing the process. Of course, it will
never be a perfect solution, as in many cases, it is required the custom
signal handler to be executed by the kernel. Still, for the moment, Tilck
cannot execute signal handlers yet.