[config] Increase KERNEL_STACK_SIZE to 8 KB
Unfortunately, with the current execve() support for scripts plus the current
stack consumption of vfs_resolve(), using just 4 KB for the kernel stack size
is not enough, in particular when trying to execute shell scripts. In that
scenario, it's likely to hit a double-fault because of stack overflow.

A good temporary solution to this problem is just to increase the stack size
to 8 KB and then to research ways to consume less stack in those two areas.