[busybox][build_toolchain] Fix (again) the busybox config file
It turned out that comparing other/busybox.config and toolchain/busybox/.config
was still not reliable on all the systems, because the locale affected the sort
command.

This patch fixes both the build_toolchain script and the saved busybox.config
file to use the simplest "C" locale while sorting. Hopefully, this is the last
fix for this problem.