[build_toolchain] Fix support for Debian on WSL
Assuming that because a distro is Debian "su" instead of "sudo" must be used
is wrong. For example, on WSL, Debian is configured to use sudo instead.

With this patch the ubuntu and the debian "install packages" code is unified.
Now, if sudo is installed, we're going to use it, otherwise we'll have to use
"su -c" instead.