[build_toolchain] Add `cmake` to all packages_list arrays
Using a local pre-built CMake version is great, but it doesn't always work.
For example, on Manjaro 18.1 (64-bit) the toolchain version of CMake fails
to start because a shared library is missing from the system.

It's much more likely to have zero problems by just using the version of CMake
part of the Linux distribution being used, than by using pre-built binary.
For this reason, with this patch, CMake becomes part of the packages that are
required by Tilck in order to compile.

In the corner case that the system has a CMake version older than 3.2, than
the toolchain CMake function will download and use the pre-built binary.

Only if not even this works, the user will have to manually deal with the
issue.