[datetime] Avoid multiplying boot_timestamp by TS_SCALE
In order to compute tv_sec, in real_time_get_timeval() it's enough to only
sum `boot_timestamp` with the scaled-down-to-second system-time.

This approach limits the max value that can be reached during the computation,
which is always a good thing to do, even if, with 64-bits, we could do the
older computation even with TS_SCALE = 10^6.