[timer] Fix a nasty bug in the wake-up timer logic
While it was nice to have tick_all_timers() not keeping the interrupts disabled
all the time, the older implementation suffered from a very subtle bug: what if
we cancel a timer (= just setting ticks_before_wake_up to 0) and then the task
gets destroyed and re-created, reusing the same heap chunk, before the next
timer tick? Well, in that case tick_all_timers() will notice ticks == 0 and
will remove the &ti->wakeup_timer_node from the wake-up list. But, since that
node has been re-initialized when the new task was created, that will cause
tick_all_timers() to loop forever at the same node, never hitting list's head.
That's unfortunate. But just making task_cancel_wakeup_timer() to remove the
task from the list wouldn't be safe either, because tick_all_timers() don't
keep the interrupts disabled all the time. That was the point. Now, it turns
out that this idea was simply bad.

Now, just reverting to a much simpler implementation w/o ticks_before_wake_up
being atomic is the right solution. Even if we keep the interrupts disabled
for some time. After all, Tilck scale (at most tens of tasks running on the
system) allows that.

After that, it would be nice to find a much better O(1) solution for
tick_all_timers() as explained in the comments. It is definively possible,
but given the constraint of not being able to use kmalloc() in
task_set_wakeup_timer(), it's not exactly trivial.