[term] Remove set_cursor_enabled() from term's interface
Now that the debug panel uses the escape sequences to do that, it's not
necessary anymore to have this function also as part of generic term's
interface (struct term_interface). The TTY layer accesses this functionality
through video term's actions, available only in filter functions. The smaller
term's interface, the better is.