[build_fatpart] Don't include the krnl in the initrd if possible
This patch actually fixes this recently-introduced feature. The bug was that
in case the bootloader files already exist because of a pre-existing build,
the script will still include the kernel in fatpart, unnecessarily.

Now the script depends on the CMake variable KERNEL_FATPART_PATH which won't
be set in case neither bootloader is used.