[sched] Disable interrupts in task_change_state()
Unfortunately, task_change_state() called directly or indirectly from many
non-IRQ handlers like kernel_yield() races with its call in tick_all_timers(),
which is called in the IRQ0 handler. While it's perfectly a good idea to
avoid disabling the interrupts as much as possible, at least changing the
state of a task must be a thing that's always safe to do. It's reasonable to
expect an IRQ handler to least to be able to change the state of a task. Sure,
a tricky solution like using a safe ringbuffer like term does it is possible
and it will avoid disabling the interrupts, but it has a significant overhead
both in terms of complexity and runtime cost. The CLI/STI instructions, on the
other side, are super-cheap and the whole task_change_state() is extremely
fast.

The race-condition this patch is going to fix shows that, despite the several
sophisticated attemps to avoid making task_change_state() to disable the
interrupts, at the end, even for just one single use-case, it turns to be
inevitable.