[config] Reduce the size of the first heap to 128 KB
As the kernel sections grow (text, data and bss), the 1 MB barrier gets closer
and closer and we're just a few KBs away from it. Now, with the new changes
in the debug panel, we're going to pass over it. In order to avoid that, we can
just shrink our balloon, the first kmalloc heap, from 256 KB to 128 KB.