[start] Always disable paging and PAE before setting CR3
In some cases, like in the scenario "qemu+kvm 32-bit efi boot" with the latest
OVMF pure UEFI firmware, we might enter in _start with paging and PAE enabled,
which is very bad. The code in start.S used to assume that could never be
possible, but obviously, it was unsafe to assume that. So, setting CR3 without
first disabling paging and PAE caused #GPF, because PGE was not enabled in CR4
and we didn't expect setting CR3 could have any effect (assuming the paging
was disabled). Now paging and PAE get always disabled before even touching CR3.