[fatpart] Use Doom1.WAD if exists in $tc/$arch/fbDOOM/
Commit ae7f2629 made possibile to use a WAD file different from the
freedoom's one. But that required the user to put it directly in
sysroot/usr/lib/fbDOOM. The problem with that approach was that the given
WAD file would be always copied in the `initrd`, not matter the value of
EXTRA_FBDOOM_ENABLED. That's not very convenient. This patch expects
the user-provided Doom1.WAD file to be present (if any) in the toolchain
instead. The `build_fatpart` will copy it instead of the freedoom one
if it exists.