[interrupts] Always run fault handlers with IF=1
Until now, soft_interrupt_entry() used to run fault handlers with interrupts
disabled, just in case was needed, even if ALL the handlers used to re-enable
the interrupts as first thing. Fault handlers are not IRQs and should not have
any reason to require entering with disabled interrupts while they might,
of course, disable the interrupts later, for a short time, as ANY other kernel
code.