[scripts] Exclude common/3rd_party from LoC counting
The purpose of the `count_lines_of_code` script is to count the amount of
physical lines of source code in the whole Tilck project, including
scripts, CMake file and any kind of infrastructure code as well.

But, the script must *not* consider generated files, documentation and
3rd party code, of course.

This simple patch just makes the script to exclude any paths containing
the `3rd_party` substring from the count.