[mbrhack] Intr. the mbrhack tool
The purpose of this tool is to clean up the mess of hard-coded calculations
for the MBR partition table in boot.S. In the following commit, boot.S won't
have anymore a hard-code partition table. It will be instead written by
this tool, called by the `tilck.img` CMake top target (see ${IMG_FILE}).

Why not using `parted` for MBR partitions?
TL;DR answer: It does not give us FULL CONTROL over the bios parameter
block (BPB) and, in general, it's not flexible enough.

For example, in order to create partitions, it requires to create its
own (standard) MBR, while we have it embedded in our boot.S file and
we're stuck with that.

    - We cannot write the `parted` MBR first and then our MBR, because
      it will overwrite it.

    - We cannot write our MBR first and then use `parted` to create
      partitions, because it doesn't like it and refuses to work.

Another solution will be to use a standard MBR (like parted one's) and then
move our boot.S at the beginning of the fatpart, hoping that the standard
MBR will chain-load the 1st sector of the FAT partition at the addr 0x07C00,
but that's not sure. Probabily, doing that will lead to other complications.
For example, we'll have to stop using `mformat` in a simple way as now to
create the FAT partition, because we'll need to add bootloader code in it.
So, it might turn out to be the MBR problem all over again.

In conclusion, it seems a good idea to have a super-powerful tool for
manipulating the MBR, having 100% control of what's happening.