[poll][select] Fix behavior for timeout < 1 timer tick
In case of timeouts smaller than the minimum possible (1 tick), the two
functions rounded it to zero and behaved exactly as if timeout = 0 was
passed to them, by returning immediately. That's a possible solution but
it's not the behavior expected by the POSIX interface. The functions should
sleep for the smallest possible time instead, 1 tick.

This patch just makes their behavior to be POSIX-compliant.