[build_fatpart] Drop EFI files and the kernel from initrd
Now that the EFI boot files and the Tilck kernel can be loaded (by both the
bootloaders) from the bootpart, we can finally remove them from the initrd.
That's great because the bootpart won't be kept in memory the whole time
like initrd does. This way, we saved about ~800 KB of (wasted) memory.