[config][cmake] Intr. the TINY_KERNEL option
This option makes the kernel as small as possible by force-disabling
some modules like 'console', 'fb', 'tracing' and by force-setting to OFF
features like KERNEL_SELFTESTS and KERNEL_SYMBOLS. In addition to that,
the option enforces constraints like CMAKE_BUILD_TYPE=Release and
KERNEL_GCOV=0. In case constraints are not met, the build simply fails in
CMake with a fatal error. Finally, the value of constants KERNEL_MAX_SIZE
and KMALLOC_FIRST_HEAP_SIZE is reduced when TINY_KERNEL=1.

This option is great because, when turned on, it allows Tilck to boot and
run perfectly fine on a VM with just 4 MB of RAM, without renouncing on
a single busybox applet.