[devshell] Restore default sig handlers when running single cmd
By default, the devshell ignores SIGINT, SIGQUIT and restores them to
their default handling before executing each command, after fork(),
in shell_run_child(). But, when devshell is run with the -c option,
a single command (typically a system test) is executed, without fork().
There was a minor bug in that case, though: the SIGINT and SIGQUIT
signals remained ignored. That was bad, in particular when running all
the test cmds with the `runall` command, simply because we couldn't just
interrupt the the whole thing with Ctrl+C.

This patch addresses the issue by moving the loop resetting all the
signal handlers from shell_run_child() to run_if_known_command().