[cmake] Intr. smart_config_file()
The purpose of this change is to drastically improve the incremental builds.
Until now, changing even minor configuration options could trigger the
rebuild of most of the Tilck project, just because the config header files
are re-generated. Note: re-generated files IDENTICAL to their previous version
still trigger re-build of the C files using them because only the timestamp
matters. That's not what we want, actually. If a configuration file didn't
change, the C files using it must NOT be re-built. To solve this problem,
we've introduced the smart_config_file() CMAKE function. It generates first
a temporary config file, compares it to the current one and, only if they
differ, the current config file is replaced by the new (temporary) one.
This *greatly* reduces the number of files that are re-built after changing
config options.