[efi] Append WARN_FLAGS_LIST to EFI_BUILD_BASE_FLAGS_LIST
This includes building the EFI bootloader with -Wshadow.
Doing that required changing bootloader's code in order to avoid shadowing
of global variables. The cases globals shadowed were just 3, repeated all over
the code: BS, ST and mbi. In all the cases the value of the parameter was
exactly the same as the global variable. Therefore, avoiding the warning meant
just removing that useless parameter. Note: gnu-efi defined BS and ST as
globals, taking their values from the InitializeLib() call in efi_main().