[i8042] Disable preemption in i8042_reboot()
Reason: in order to request the system reboot, i8042_reboot() calls
i8042_send_cmd() which calls indirectly delay_us() through i8042_io_wait().
The problem is that delay_us() ASSERTs that the preemption is disabled.
The bug is that, currently, during reboot the preemption is kept enabled
(while the interrupts are disabled, of course). Therefore, the reboot ends
up causing a kernel panic due to the failed ASSERT. This patch fixes this
simple issue by disabling the preemption at the beginning of i8042_reboot().