[scripts] Intr. the scripts/dev/run_config script
This script is just a wrapper of ccmake with some nice error checking.
Also, in this patch the `config` and `menuconfig` targets have been added
to the main (wrapper) Makefile. In order words, running:

	make menuconfig

From project's root directory, will run this script and ultimately run ccmake.
In the future, there might be a real menuconfig target with a curses GUI nicer
than ccmake, but for the moment we're stuck with it.