[sched][timer] Don't ASSERT state == sleeping in the timer_ready case
In case a task in kernel_sleep() is preempted before kernel_yield() is
called, it is possible the timer to fire much before the task is resumed.
When that happens, if the task in kernel_sleep() didn't get a chance to
change its state to SLEEPING yet, the timer won't wake-up the task, because
it's not sleeping. After that, when the task is resumed, it will go to sleep
forever (the timer already fired). This race condition has ALREADY been
handled in schedule() with the help of the per-task flag called timer_ready
BUT the fix had a subtle bug: it ASSERTed that current task's state is
SLEEPING if it's not RUNNING. That implied that the race would occur only
between task_set_wakeup_timer() and task_change_state(). Of course, it was
clear at the time that the race could also occur immediately before the final
call to kernel_yield(), but that seemed a perfectly harmless case. The reality
is that the case is indeed harmless, but schedule's ASSERT did not consider
that when the timer fires before kernel_yield(), it also sets the "SLEEPING"
task in the RUNNABLE state, which is neither RUNNING nor SLEEPING.

In conclusion, the whole fix for this tricky race condition is just to drop
the wrong ASSERT from schedule() in the timer_ready case.