[sched] Handle enter_sleep_wait_state() racing with a wake-up timer
This patch fixes a race similar to the one described in the previous commit.
In this case, in functions like poll() or select() (where the bug has hit),
first a wake-up timer is set, then prepare_to_wait_on() is called (directly
or indirectly) and then enter_sleep_wait_state() is called. The problem was
that enter_sleep_wait_state() used to ASSERT that current task's state is
SLEEPING, while in case the timer fired before getting there, task's state
would be RUNNABLE. While thanks to the previous commit, schedule() handles
this case, in other cases we can't get there because enter_sleep_wait_state()
ASSERTs that the task's state is SLEEPING.

The current patch relaxes a little bit enter_sleep_wait_state()'s expectations
and allows it to get to kernel_yield() even if task's state is not SLEEPING as
long as the timer_ready flag is set.