[execve][elf] Handle differently wrong-arch ELF binaries
Tilck has no problem to just return -ENOEXEC in this case, but
ASH's implementation in busybox does not behave as it should: it
tries to interpret ELF's content AS IF it were a shell script: that
causes weird failures, not affecting Tilck's kernel but still, it
makes harder to realize that the ELF had just the wrong arch (e.g.
compiled for x86_64 instead of i686).