[execve][elf] Handle with an error the case of dynamic executables
Since Tilck has no support for dynamic executables, in case a dynamic
executable is run, Tilck will ignore the INTERP segment and it will
jump directly to ELF's entry point, instead of jumping to dynamic linker's
entry point. Therefore, at the first access to a symbol not in the ELF file
itself we'll get a SIGSEGV.

In order to avoid wasting people's time debugging why a given program
crashes on Tilck, but not on Linux, it's nice to fail early and display
a meaningful message. That's what this patch does.