[toolchain] Don't run `make clean` in build_lua()
The build_XXX funcs part of the build_toolchain script are designed
to be idempotent, which means that calling them multiple times has
no effect, after the first time. Because of that, rebuilding any given
app/package in the toolchain is not supported: it can only be forced
after manually removing package's directory (e.g. toolchain2/i386/lua).
This patch just removes the `make clean` command before the first and
only build of lua, simply because it's not needed and it has the side
effect of displaying a very long `rm -rf ...` message the first time
lua is built.