[process] Set vforked=false in allocate_new_process()
When new_pdir == parent_pi->pdir, the function sets pi->vforked = true
but, in the other case, the `vforked` value of its parent is kept (simply
because all the fields are copied with memcpy()). That's not correct:
a vforked process can indeed call fork() and the new process will be a
regular process, not a vforked one. That use-case can be triggered by the
`timeout` busybox applet this way:

	timeout -t 1 sleep 2

This one-line patch sets pi->vforked=false in the non-vfork case.
Without it, when we run the timeout applet, we end up triggering an
ASSERT(parent->stopped) in the exit code path. That's because pi->vforked
is (wrongly) set and therefore we have wrong assumptions about process'
parent. From another point of view, the ASSERT really helped finding the
real problem in an easier way, because of the early-fail.