[kernel] Merge all .ctors.* sections into .ctors
With GCC >= 7.0, when -fprofile-arcs is passed as option, the per-TU ctors
are placed in .ctors.NNN (where NNN is just a number) instead of in .ctors
as with older compilers. That made Tilck's call_kernel_global_ctors() to stop
calling those ctors simply because the function just calls the function
pointers in the ".ctors" section and doesn't know about the other one.

To avoid this problem, kernel's linker script has been updated to merge
any .ctors.* sections in one single section called ".ctors".