[lcov][toolchain] Upgrade lcov to version 1.15
Unfortunately, the lcov package need to be updated in order to keep working
with GCDA files generated by GCC >= 9.0. This change introduces a
LCOV_VER="1.15" variable both in the CMake files (for error-checking)
and in the `scripts/bash_includes/tc/lcov` script as well. In addition to
that, now lcov is downloaded to "$TC/noarch/lcov-$LCOV_VER". In case an
old `lcov` directory is found there, it gets renamed to lcov-1.13 and a "lcov"
symlink is created to it.