[runners] Raise MED_TEST_TIMEOUT from 24 to 36 seconds
There is strong evidence that the `kmutex_med` self test failed at least once
because the timeout was too small. During that test run, the deadlock detection
logic continued to report strong progress towards the 100k iterations goal,
but at about ~97,800 iterations (very close for all the 3 threads), the runner
killed the VM because of the timeout. Therefore, that was not a real failure
and only right fix is to just raise the timeout. That's what this patch does.