[system_tests] Intr. a simple script-test for sysfs
While sysfs deserves proper and extensive testing through a dedicated
unit test, doing that requires a fair amount of time and effort and that's
not affordable right now, while certainly it's in a TODO list.

Adding this simple test script that uses sysfs certainly is not enough to
test sysfs, but it's just better than nothing. It triggers many code paths
that weren't triggered before by tests and, in case of something brutally
broken with sysfs, it will fail indeed.